[FEAT] Finalize the project
- Delete experimental part of dictionary
- Separate anagram search and results in different tabs
