+++ backend/.env
@@ -1,7 +1,3 @@
 # ANAGRAM
 ANAGRAM_INPUT_TEXT_MAX_LENGTH="10"
 ANAGRAM_INPUT_VALIDATION_REGEX="[^a-zA-Z]"
-
-# DICTIONARY
-DICTIONARY_AVAILABLE_LANGUAGES="it" # Use pipe (|) character to separate
-DICTIONARY_INPUT_WORD_MAX_LENGTH="30"
