--- backend/__tests__/api/dictionary/dictionary.test.type.ts
@@ -1,4 +1,0 @@
-export interface DictionaryResposneType {
-  statusCode?: number;
-  text?: string;
-}
