--- frontend/src/components/anagrams-tab-anagram/anagrams-tab-anagram.type.ts
@@ -1,10 +1,0 @@
-export interface AnagramsTabAnagramType {
-  text: string;
-  setText: (text: string) => void;
-}
-
-export interface AnagramsReturnMessage {
-  title?: string;
-  detail?: string;
-  data?: string[];
-}
