+++ frontend/src/components/anagrams-tab-search/anagrams-tab-search.type.ts
@@ -1,0 +1,10 @@
+export interface AnagramsTabAnagramType {
+  text: string;
+  setText: (text: string) => void;
+}
+
+export interface AnagramsReturnMessage {
+  title?: string;
+  detail?: string;
+  data?: string[];
+}
