+++ frontend/src/components/anagrams-tabs/anagrams-tabs.lib.ts
@@ -1,6 +1,7 @@
-import { BookOpenIcon, Squares2X2Icon } from '@heroicons/react/24/outline';
+import { Squares2X2Icon } from '@heroicons/react/24/outline';
 
 import { TabsType } from './anagrams-tabs.type';
+
 
 /**
  * @function getTabs
@@ -8,6 +9,6 @@ import { TabsType } from './anagrams-tabs.type';
  * @returns {TabsType} Tabs details
  */
 export const getTabs = (): TabsType[] => [
-  { id: 'anagrams', name: 'Anagrams', icon: Squares2X2Icon },
-  { id: 'dictionary', name: 'Dictionary', icon: BookOpenIcon },
+  { id: 'search', name: 'Find Anagrams', icon: Squares2X2Icon },
+  { id: 'results', name: 'Results', icon: Squares2X2Icon },
 ];
