--- frontend/src/redux/reducers/dictionary/dictionarySlice.type.ts
@@ -1,5 +1,0 @@
-export interface DictionaryStateType {
-  found: boolean;
-  status: string;
-  error: string;
-}
