+++ frontend/src/redux/reducers/tabs/tabsReducers.ts
@@ -1,0 +1,11 @@
+import { PayloadAction } from '@reduxjs/toolkit';
+
+import { tabsIdType, TabsStateType } from './tabsSlice.type';
+
+
+/**
+ * REDUCER: setText
+ */
+export const reduceSetCurrentTab = (state: TabsStateType, action: PayloadAction<tabsIdType>) => {
+  state.currentTab = action.payload;
+};
