+++ frontend/src/redux/reducers/tabs/tabsSlice.type.ts
@@ -1,0 +1,5 @@
+export type tabsIdType = 'results' | 'search';
+
+export interface TabsStateType {
+  currentTab: tabsIdType;
+}
