+++ frontend/src/redux/store.ts
@@ -1,10 +1,13 @@
 import { configureStore } from '@reduxjs/toolkit';
 
 import anagramReducer from './reducers/anagram/anagramSlice';
+import tabsReducer from './reducers/tabs/tabsSlice';
+
 
 const store = configureStore({
   reducer: {
     anagram: anagramReducer,
+    tab: tabsReducer,
   },
 });
